requirements.txt
anthropic>=0.40.0
instaloader>=4.13
python-dotenv>=1.0.0
